fix compilation error for the DPI fix

+++ src/window.cxx
@@ -43,7 +43,7 @@ namespace Moon {
 
     // Sets up the projection matrix so that (0,0) corresponds to the top left corner,
     // and (width, height) corresponds to the bottom right corner.
-    Shader::projection_matrix = glm::ortho(0.f, width, height, 0.f, -1.f, 1.f);
+    Shader::projection_matrix = glm::ortho(0.f, (float)width, (float)height, 0.f, -1.f, 1.f);
   }
 
   bool Window::should_close() {
